Add support for Japanese characters (#1005)
This took me A LOT of time, mainly because I am not aware of any good
dictionary/map implementation in C. One of the time consuming parts was
adding the special characters due to how the game uses the dakuten and
handakuten.

The function `AnsiToSotnMenuString` is still very fast as I am searching
the characters via a binary search. If you have any suggestion for a
O(1) implementation please let me know.

This fixes some buffer overflows when rendering Japanese strings.