Use player step enum in EntityAlucard (#1108)
While attempting to trace the process of an action in the game, I
noticed that the main state machine in EntityAlucard which dispatches
the functions for each player state was still using numerical `case`
statements. I cross-referenced the `PlayerSteps` enum in `dra.h` to the
steps in the function, and swapped in all the enum names. I also gave
proper names to unk24 and unk25, deconflicting this in the Alucard and
Richter player step lists.

Finally, I noticed that EntityAlucard has routines for steps 48-50,
which did not have enum members, so I added those. We'll have to examine
what they do.