Decompile weapon 16 EntityWeaponAttack (#1129)
This is mostly just another weapon function.

However, because this references PLAYER.step, we want to use the
PlayerSteps enum. This enum is in `dra.h` and therefore inaccessible to
weapon functions.

I ran into this problem a while ago dealing with ric.h not having access
to the steps, and just duplicated PlayerSteps. Now that we have a third
place, we instead move PlayerSteps to `game.h`, and adjust `ric.h` to
only use the steps that we need to override. This seems like the most
sustainable solution.