Decompile weapon 37 helper 8017ADF8 (#1134)
Not really certain what this one does. Given that weapon 37 is for the
short sword and jewel sword, I wonder if it's related to the colorful
particles that come out when you swing the jewel sword. I'm sure we'll
learn more as we decompile the other functions in the overlay.