Shared `CollectHeartVessel` Implementation
This follows the model of [/src/st/collect_hearts.h] to reuse the
`CollectHeartVessel` implementation across stages that use the same
implementation.

This adds implementations for CEN and WRP as well.

One concern I have about this approach is that header dependencies
aren't mapped well in the Makefiles at the moment, so changes to these
file will not trigger a recompile.