TT_000 PSP func_801713C8 (#1143)
Matches `func_801713C8`.

This is an important milestone as it is the first function using the
`SEH` instruction. I got this new compiler (and many more) from someone
in ObscureGamers that kindly shared at
https://archive.org/download/compilers-no-license .

This still gives us no clue on what `f32` and the various `.i.hi` stuff
was actually meant to be.