Decompile weapon 41 helper 8017A994 (#1156)
Very neat to solve this with the help of the PSP version! I couldn't
solve this without it, so this is a cool development. Hoping to keep
benefiting from that!

This is also the final helper function in the weapons. All other
functions in all the weapon overlays are part of the main Weapon struct
that holds the set of several pointers to call.