Decompile MAD CreateEntitiesToTheLeft
Every other stage uses the same implementation of
`CreateEntitiesToTheLeft`, but MAD's version does not loop through all
entities which may be in range, it finds the first, and depending on
flags may create it, the array pointer is updated and no other entities
are considered. This function is used when scrolling left, meaning if
multiple entities fall within the scroll delta, only the first will be
created which might have some interesting properties.