Decompile DRE Entity3DHouseSpawner (#1260)
This was an old one with a decompme comment in the code that I decided
to tackle. Turned out to be able to make it relatively straightforward!

I think the naming is reasonable and is motivated by the behavior in the
code, but happy to adjust if needed.

Strangely, decompme informed me that the function was matched here:
https://decomp.me/scratch/aEXJO by "NicoVazCan", 3 months ago. Not sure
who that is, but they never submitted it, so I guess it's abandoned.
Either way, I think the code in this PR is more readable.